[SettingsForm] test whether or not directories exist for backup_CheckedChanged() event.
git-svn-id: http://svn.code.sf.net/p/megui/code@1210 a21b8521-4df1-4ed3-b48c-cff5bc73dcb3

+++ megui/trunk/core/gui/SettingsForm.cs
@@ -2487,17 +2487,30 @@ namespace MeGUI
             {
                 string meguiToolsFolder = Application.ExecutablePath.Substring(0, Application.ExecutablePath.LastIndexOf('\\')) + "\\tools\\";
                 string meguiAvisynthFolder = MeGUISettings.AvisynthPluginsPath + "\\";
-                try
-                {  // remove all backup files found
-                    Array.ForEach(Directory.GetFiles(meguiToolsFolder, "*.backup", SearchOption.AllDirectories),
-                      delegate(string path) { File.Delete(path); });
-                    Array.ForEach(Directory.GetFiles(meguiAvisynthFolder, "*.backup", SearchOption.AllDirectories),
-                      delegate(string path) { File.Delete(path); });
-                }
-                catch (Exception ex)
+                if (Directory.Exists(meguiToolsFolder))
                 {
-                    MessageBox.Show(ex.Message, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                } 
+                    try
+                    {  // remove all backup files found
+                        Array.ForEach(Directory.GetFiles(meguiToolsFolder, "*.backup", SearchOption.AllDirectories),
+                          delegate(string path) { File.Delete(path); });
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show(ex.Message, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                }
+                if (Directory.Exists(meguiAvisynthFolder))
+                {
+                    try
+                    {  // remove all backup files found
+                        Array.ForEach(Directory.GetFiles(meguiAvisynthFolder, "*.backup", SearchOption.AllDirectories),
+                          delegate(string path) { File.Delete(path); });
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show(ex.Message, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                }
             }
         }
